Refactor E2E login: unify login logic, update selectors, move and rename specs
