+++ frontend/tests/e2e/login/specs/login-redirect.spec.ts
@@ -1,7 +1,7 @@
 import { expect, Page, test } from "@playwright/test";
 import { VALID_TAGS } from "tests/e2e/tags";
 
-import playwrightEnv from "./playwright-env";
+import playwrightEnv from "tests/e2e/playwright-env";
 
 const { AUTH, SMOKE, CORE_REGRESSION, FULL_REGRESSION } = VALID_TAGS;
 const { baseUrl, targetEnv } = playwrightEnv;
@@ -32,7 +32,7 @@ const setupLoginRedirectSpoof = async (page: Page) => {
       await route.fulfill({
         status: 200,
         contentType: "text/html",
-        body: `<html><head><meta http-equiv="refresh" content="0;url=${baseUrl}/"></head></html>`,
+        body: `<html><head><meta http-equiv=\"refresh\" content=\"0;url=${baseUrl}/\"></head></html>`,
       });
     } else {
       await route.continue();
