+++ frontend/tests/e2e/playwrightUtils.ts
@@ -1,9 +1,4 @@
-import {
-  BrowserContextOptions,
-  expect,
-  FullProject,
-  Page,
-} from "@playwright/test";
+import { BrowserContextOptions, expect, Page } from "@playwright/test";
 import playwrightEnv from "tests/e2e/playwright-env";
 
 const { targetEnv } = playwrightEnv;
@@ -182,46 +177,6 @@ export const generateRandomString = (desiredPattern: number[]) => {
     }
     return randomString;
   }, "");
-};
-
-/* ---- Auth /Sign In ---- */
-
-// signs in using mock 0auth server
-// note that this does not currently work in CI, but does work locally
-// an unknown error prevents sending the token back successfully from the API in CI
-// this will be remedied by https://github.com/HHS/simpler-grants-gov/issues/3791
-// after which we can reenable sign in related tests
-export const performSignIn = async (page: Page, project: FullProject) => {
-  const signInButton = page.locator('button[data-testid="sign-in-button"]');
-  await expect(signInButton).toHaveText("Sign in");
-  await signInButton.click();
-  const secondSignInButton = page.locator(".usa-modal__footer a");
-  await secondSignInButton.click();
-
-  await waitForAnyURLChange(page, "/");
-
-  const requiredInput = page.locator('input[type="text"]');
-  const submitButton = page.locator('input[type="submit"]');
-  const randomUserName = generateRandomString([12]);
-
-  await requiredInput.fill(randomUserName);
-  await submitButton.click();
-
-  await waitForUrl(page, "http://localhost:3000/");
-
-  if (project.name.match(/[Mm]obile/)) {
-    const userDropdown = page.locator(
-      'button[data-testid="navDropDownButton"]',
-    );
-    await userDropdown.click();
-    await expect(
-      page.locator("#user-control > li:first-child a div"),
-    ).toHaveText("fake_mail@mail.com");
-  } else {
-    await expect(
-      page.locator('button[data-testid="navDropDownButton"] a div'),
-    ).toHaveText("fake_mail@mail.com");
-  }
 };
 
 /* ---- Mobile Navigation ---- */
